requirements.txt
#-------------------------------------------------------------------------------
# Django
#-------------------------------------------------------------------------------

Django==1.6.2

#-------------------------------------------------------------------------------
# other packages
#-------------------------------------------------------------------------------

django-debug-toolbar==1.0.1
django-haystack==2.1.0
django-simple-captcha==0.4.1
Jinja2==2.7.2
Pillow==2.3.0
pip==1.5.4
psycopg2==2.5.2
py-bcrypt==0.4
pyelasticsearch==0.6.1
python3-memcached==1.51
pytz==2013.9
setuptools==2.2
South==0.8.4
Unidecode==0.04.14
watchdog==0.7.1

#-------------------------------------------------------------------------------
# deployment
#-------------------------------------------------------------------------------

uWSGI==2.0.2
